+++ src/Modules/Message/Leelite.Modules.MessageCenter/Repositories/PushPlatformRepository.cs
@@ -5,11 +5,11 @@ using Leelite.Modules.MessageCenter.Models.PlatformAgg;
 
 namespace Leelite.Modules.MessageCenter.Repositories
 {
-    public interface IPushPlatformRepository : IRepository<PushPlatform, long>
+    public interface IPushPlatformRepository : IRepository<PushPlatform, int>
     {
     }
 
-    public class PushPlatformRepository : EFRepository<PushPlatform, long>, IPushPlatformRepository
+    public class PushPlatformRepository : EFRepository<PushPlatform, int>, IPushPlatformRepository
     {
         public PushPlatformRepository(MessageContext dbContext, IUnitOfWork unitOfWork) : base(dbContext, unitOfWork) { }
     }
